Plugin testing infrastructure added to local Bloonix::Plugin.
This should exist only in the repository and not released. It adds
a --test <filename> option to Bloonix::Plugin so plugins can be
tested in isolation without actually making the requests to external
or local services.

The test file contains test information that would have been returned
by the external or local service so the rest of the plugin behaviour
can be tested and multiple conditions can be tested.